add button and use dataset name for plot title

+++ interface_app.py
@@ -13,12 +13,11 @@ from tagging import data, plot
 
 data_path = pathlib.Path("/tmp")
 upload_folder_name = "upload"
-dataset_name = st.text_input("Enter the name of your dataset", value="Test")
 
+dataset_name = st.text_input("Enter the name of your dataset", value="Test")
 naming_scheme = st.radio("Naming Scheme for phases", ["old", "new"])
 naming_suffix = st.text_input("Suffix for relevant phase names", value="_peak")
 uploaded_file = st.file_uploader("Choose a CSV file", type="csv")
-
 
 home_possession_name = "possesion"
 away_counter_name = "negative_transition"
@@ -31,19 +30,21 @@ if naming_scheme == "new":
     home_counter_name = "home_counter"
         
 if uploaded_file is not None:
-    data.extract_single_csv(uploaded_file, data_path, file_from_disk=False)    
-    fig, ax = plot.make_phase_plot_for_dataset(
-        data_path, 
-        upload_folder_name, 
-        home_possession_name,
-        away_counter_name,
-        away_possession_name,
-        home_counter_name,
-        file_suffix="_peak"
-    )
+    data.extract_single_csv(uploaded_file, data_path, file_from_disk=False)
     
-    shutil.rmtree(data_path / upload_folder_name)
-
-    st.pyplot(fig)
+    if st.button("Create graphs"):
+        fig, ax = plot.make_phase_plot_for_dataset(
+            data_path, 
+            upload_folder_name, 
+            home_possession_name,
+            away_counter_name,
+            away_possession_name,
+            home_counter_name,
+            title=dataset_name,
+            file_suffix=naming_suffix,
+        )
+    
+        shutil.rmtree(data_path / upload_folder_name)
+        st.pyplot(fig)
